chore(catalog): Refactor schema/types
