+++ internal/provider/catalog/catalog_schema.go
@@ -18,38 +18,6 @@ import (
 	superschema "github.com/FrangipaneTeam/terraform-plugin-framework-superschema"
 )
 
-type catalogDataSourceModel struct {
-	// BASE
-	ID          types.String `tfsdk:"id"`
-	Name        types.String `tfsdk:"name"`
-	Description types.String `tfsdk:"description"`
-	CreatedAt   types.String `tfsdk:"created_at"`
-	OwnerName   types.String `tfsdk:"owner_name"`
-
-	// SPECIFIC DATA SOURCE
-	PreserveIdentityInformation types.Bool  `tfsdk:"preserve_identity_information"`
-	NumberOfMedia               types.Int64 `tfsdk:"number_of_media"`
-	MediaItemList               types.List  `tfsdk:"media_item_list"`
-	IsShared                    types.Bool  `tfsdk:"is_shared"`
-	IsPublished                 types.Bool  `tfsdk:"is_published"`
-	IsLocal                     types.Bool  `tfsdk:"is_local"`
-	IsCached                    types.Bool  `tfsdk:"is_cached"`
-}
-
-type catalogResourceModel struct {
-	// BASE
-	ID          types.String `tfsdk:"id"`
-	Name        types.String `tfsdk:"name"`
-	Description types.String `tfsdk:"description"`
-	CreatedAt   types.String `tfsdk:"created_at"`
-	OwnerName   types.String `tfsdk:"owner_name"`
-
-	// SPECIFIC RESOURCE
-	StorageProfile  types.String `tfsdk:"storage_profile"`
-	DeleteForce     types.Bool   `tfsdk:"delete_force"`
-	DeleteRecursive types.Bool   `tfsdk:"delete_recursive"`
-}
-
 func catalogDatasourceAttributes() map[string]schemaD.Attribute {
 	attr := catalogSchema().GetDataSource(context.Background()).Attributes
