+++ internal/provider/catalog/catalogs_types.go
@@ -1,0 +1,11 @@
+package catalog
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+type catalogsDataSourceModel struct {
+	ID           types.String                      `tfsdk:"id"`
+	Catalogs     map[string]catalogDataSourceModel `tfsdk:"catalogs"`
+	CatalogsName types.List                        `tfsdk:"catalogs_name"`
+}
